added authorization to update and delete

+++ backend/controllers/taskController.js
@@ -32,7 +32,6 @@ const getTasks = async (req, res) => {
             return res.status(404).json({ message: 'No tasks found' });
         }
 
-
         res.json(tasks);
 
     } catch (err) {
@@ -48,6 +47,11 @@ const getTaskById = async (req, res) => {
         if (!taskById) {
             return res.status(404).json({ error: 'Task not found' });
         }
+
+        if (taskById.user.toString() !== req.user.id) {
+            return res.status(403).json({message: 'Not authorized'})
+        }
+        
         res.json(taskById);
     } catch (err) {
         res.status(500).json({ error: 'Server error' });
@@ -75,16 +79,24 @@ const createTask = async (req, res) => {
 const updateTask = async (req, res) => {
     try {
         const { id } = req.params;
+        const task = await Task.findById(id);
+
+        if (!task) {
+            return res.status(404).json({message: 'Task not found'})
+        }
+
+        if(task.user.toString() !== req.user.id) {
+            return res.status(403).json({message: 'Not authorized'});
+        }
+
         const updatedTask = await Task.findByIdAndUpdate(
             id,
             req.body,
             { new: true }
         );
-
-        if (!updatedTask) {
-            return res.status(404).json({ error: 'Task not found' });
-        }
+        
         res.json(updatedTask);
+
     } catch (err) {
         res.status(500).json({ error: 'Server error' });
     };
@@ -94,11 +106,18 @@ const updateTask = async (req, res) => {
 const deleteTask = async (req, res) => {
     try {
         const { id } = req.params;
-        const deletedTask = await Task.findByIdAndDelete(id);
 
-        if (!deletedTask) {
-            return res.status(404).json({ error: 'Task not found' });
+        const task = await Task.findById(id);
+
+        if (!task) {
+            return res.status(404).json({message: 'Task not found'})
         }
+
+        if(task.user.toString() !== req.user.id) {
+            return res.status(403).json({message: 'Not authorized'});
+        }
+        await Task.findByIdAndDelete(id);
+
         res.json({ message: 'Task Deleted Successfully' });
 
     } catch (err) {
